13. cdp_lldp_neighbor_map/requirements.txt
netmiko
pyyaml
networkx
matplotlib
